Copy the code for encoding VP9 over from Software renderer
